// types/Remarks.tsx
export type FinalRemarks = {
  very_good: {
    message: string[];
    margin: number | null;
  };
  good: {
    message: string[];
    margin: number | null;
  };
  average: {
    message: string[];
    margin: number | null;
  };
  poor: {
    message: string[];
    margin: number | null;
  };
  very_poor: {
    message: string[];
    margin: number | null;
  };
};